Made python-dependent plugins optional
Biggest hangup to using these files on Windows. Moved the plugins that
require Python(3) to an opt pach and added a little loader in `/plugins`

Moved LaTeX to opt as well because it loads a lot no matter what kind of
file you're editing.

Switched rspec plugin to one that works

Removed nodejs: it just looks for the shebang and sets the filetype to
javascript - something that the javascript plugin _also_ does

Removed css-color: it only works on vanilla css, but runs a function all
the time